requirements.txt
streamlit==1.38.0
pyodbc==5.1.0
google-generativeai==0.7.2
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.2
seaborn==0.13.2
langchain==0.2.16
langchain-google-genai==1.0.10
plotly==5.18.0

